+++ src/libxtp/gwbse/sigma_base.cc
@@ -28,7 +28,6 @@ namespace xtp {
 
 Eigen::MatrixXd Sigma_base::CalcExchangeMatrix() const {
   Eigen::MatrixXd result = Eigen::MatrixXd::Zero(_qptotal, _qptotal);
-  Index auxsize = _Mmn.auxsize();
   Index occlevel = _opt.homo - _opt.rpamin + 1;
   Index qpmin = _opt.qpmin - _opt.rpamin;
 #pragma omp parallel for schedule(dynamic)
@@ -36,13 +35,12 @@ Eigen::MatrixXd Sigma_base::CalcExchangeMatrix() const {
     const Eigen::MatrixXd& Mmn1 = _Mmn[gw_level1 + qpmin];
     for (Index gw_level2 = gw_level1; gw_level2 < _qptotal; gw_level2++) {
       const Eigen::MatrixXd& Mmn2 = _Mmn[gw_level2 + qpmin];
-      double sigma_x = -(Mmn1.block(0, 0, occlevel, auxsize)
-                             .cwiseProduct(Mmn2.block(0, 0, occlevel, auxsize)))
-                            .sum();
-      result(gw_level1, gw_level2) = sigma_x;
+      double sigma_x =
+          -(Mmn1.topRows(occlevel).cwiseProduct(Mmn2.topRows(occlevel))).sum();
       result(gw_level2, gw_level1) = sigma_x;
     }
   }
+  result = result.selfadjointView<Eigen::Lower>();
   return result;
 }
 
@@ -51,9 +49,8 @@ Eigen::VectorXd Sigma_base::CalcCorrelationDiag(
   Eigen::VectorXd result = Eigen::VectorXd::Zero(_qptotal);
 #pragma omp parallel for schedule(dynamic)
   for (Index gw_level = 0; gw_level < _qptotal; gw_level++) {
-    double sigma_c =
+    result(gw_level) =
         CalcCorrelationDiagElement(gw_level, frequencies[gw_level]);
-    result(gw_level) = sigma_c;
   }
   return result;
 }
@@ -66,10 +63,10 @@ Eigen::MatrixXd Sigma_base::CalcCorrelationOffDiag(
     for (Index gw_level2 = gw_level1 + 1; gw_level2 < _qptotal; gw_level2++) {
       double sigma_c = CalcCorrelationOffDiagElement(
           gw_level1, gw_level2, frequencies[gw_level1], frequencies[gw_level2]);
-      result(gw_level1, gw_level2) = sigma_c;
       result(gw_level2, gw_level1) = sigma_c;
     }
   }
+  result = result.selfadjointView<Eigen::Lower>();
   return result;
 }