+++ src/libxtp/gyration.cc
@@ -20,7 +20,7 @@
 #include <boost/format.hpp>
 #include <votca/tools/elements.h>
 #include <votca/xtp/gyration.h>
-
+#include <votca/xtp/vxc_grid.h>
 using namespace std;
 using namespace votca::tools;
 
@@ -44,7 +44,6 @@ void Density2Gyration::AnalyzeDensity(const Orbitals& orbitals) {
                             << " threads ===== " << flush;
 
   const QMMolecule& Atomlist = orbitals.QMAtoms();
-  Eigen::MatrixXd DMAT_tot;
   BasisSet bs;
   bs.Load(orbitals.getDFTbasisName());
   AOBasis basis;
@@ -52,11 +51,12 @@ void Density2Gyration::AnalyzeDensity(const Orbitals& orbitals) {
   AnalyzeGeometry(Atomlist);
 
   // setup numerical integration grid
-  NumericalIntegration numway;
-  numway.GridSetup(_gridsize, Atomlist, basis);
+  Vxc_Grid grid;
+  grid.GridSetup(_gridsize, Atomlist, basis);
+  DensityIntegration<Vxc_Grid> numway(grid);
 
   if (!_dostateonly) {
-    Eigen::MatrixXd DMATGS = orbitals.DensityMatrixFull(_state);
+    Eigen::MatrixXd DMAT_tot = orbitals.DensityMatrixFull(_state);
     Gyrationtensor gyro = numway.IntegrateGyrationTensor(DMAT_tot);
     Eigen::SelfAdjointEigenSolver<Eigen::Matrix3d> es;
     es.computeDirect(gyro.gyration);